requirements.txt
boto3>=1.28.0
python-dotenv>=0.21.0
requests>=2.31.0
numpy>=1.21.0
pysqlite3-binary>=0.5.0
